Add enable_cors flag
As outlined here: https://github.com/Miserlou/Zappa/issues/494 setting
up a Zappa application to use CORS along with an api-gateway managed
authentication mechanism requires enabling cors at the api-gateway layer
to deal with the pre-flight options requests. There is a fairly simple
button to manage this in the API-gateway management console, but I'd
rather not have to take those steps after doing a deployment.

This commit gets it working with simple defaults for some of the headers.
Before I submit the pull request I want to extend it to support more fine
tuning of the available options.