Cache and reuse PyPI metadata of packages
This commit is a fix for https://github.com/Miserlou/Zappa/issues/899

The zappa package command is slower than ideal. One reason is that Zappa
checks whether a package has a manylinux wheel by downloading the package's
Python metadata from PyPI. When a project's virtualenv has a lot of
packages, this can consume a lot of time.

This commit introduces a cache for PyPI metadata and uses them in future
invocations.