Fixed when running zappa with no parameters
Fixed issue when running Zappa with no arguments. First, there was no
checking to verify if any command was passed as an argument in the
command-line and later on the dispatch_command function we were trying
to get the value of 'json' in self.vargs causing an unhandled KeyError
exception. Now, if no argument is specified, the help will be
displayed to the users.