Move task decorator async decision to call time.
Move the `@task` decorator's decision about whether to run the wrapped
function immediately or dispatch it asynchronously from import time to call
time. Previously the `lambda_function_name` was being fetched from the
environment when the to-be-async function was being imported, making testing
the async nature of the function more difficult in a test setup where the
module was already imported.

During testing it can be useful to ensure that the application under test
performs when tasks are executed either blocking or asynchronously. Although
it could be argued that if we are testing that a third party library does the
right thing we are testing the wrong thing, or that we should do integration
testing in a docker-like environment. But I thought this change couldn't hurt
anyway.